+++ airbyte-webapp/src/components/JobItem/components/ErrorDetails.tsx
